+++ bin/static/app.js
@@ -2344,4 +2344,4 @@
   };
 })();
 /* ui-bundle: f2b3a730fd0c60b7f82529bfdc699a0378f1581862d21576136480f73b00a0ba */
-/* ui-sources: e43c7b5f5bfc8acefb2a4926ed5abb1d80d67d1d6858d92a832c39899318198a */
+/* ui-sources: d04cc1c444cb6fb4c665caa210c3063f5582c1a832dfcb7bf4a451bbe0e3d26a */